task/web/requirements.txt
Django==3.1.1
djangorestframework==3.11.1
gunicorn==19.9.0


